Add required fee to submit adoption application
Add the ability to charge a fee to complete the adoption application via
the online form. The amount will be stored in a database system setting
with an initial amount of $25.00.

Required changes

Add system setting which can be modified from the Admin portal to adjust
the amount charged on the adoption application
Initialize value in database to $25.00
Add payment and billing information to the bottom of the form similar to
the donation page
Add amount paid to the confirmation email sent to the adopter and
contact@ when the application is submitted
Add comments section to website to match form
update adoption application path to correct filename spelling error
update adoption application issues with filter check boxes have
incorrect options